Fix code deprecations (#8)

+++ src/main/scala/net/soundmining/synth/Utils.scala
@@ -1,5 +1,5 @@
 package net.soundmining.synth
 
 object Utils {
-  def absoluteTimeToMillis(time: Double): Long = (time * 1000).round.toLong
+  def absoluteTimeToMillis(time: Double): Long = (time * 1000).round
 }
